services/fastapi_quote_api/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
psycopg[binary]>=3.2,<4.0
python-multipart>=0.0.9,<1.0
python-dotenv>=1.0,<2.0
